agregadas principales validaciones y asociaciones

+++ app/controllers/chofers_controller.rb
@@ -1,22 +1,22 @@
 class ChofersController < ApplicationController
 		before_action :authenticate_user! #, except: [:show, :index]
+		before_action :set_chofer, only: [:show, :edit, :update, :destroy]
 
 	def index
-		@chofers = Chofer.all()
+		@chofer = Chofer.all
 	def new
-		@chofer = Chofer.new()
+		@chofer = Chofer.new
 	end
 	def show
 		@chofer = Chofer.find(params[:id])
 	end
 	def create
-		parametros = params.require(:chofer).permit(:nombre, :apellido, :dni, :inicio_actividad)	
-		@chofer = Chofer.new(parametros)
+		@chofer = Chofer.new(chofer_params)
 	  	respond_to do |format|
 	  		if @chofer.save
 	  			format.html { redirect_to chofers_path, notice: 'El chofer se actualizo correctamente.' }
 	  		else
-	  			format.html { render :index }
+	  			format.html { render :new }
 	  		end
 	  	end
   	end
@@ -26,9 +26,8 @@ class ChofersController < ApplicationController
 
     def update
     	@chofer = Chofer.find(params[:id])
-    	parametros = params.require(:chofer).permit(:nombre, :apellido, :dni, :inicio_actividad)
 	  	respond_to do |format|
-	  		if @chofer.update(parametros)
+	  		if @chofer.update(chofer_params)
 	  			format.html { redirect_to chofers_path, notice: 'El chofer se actualizo correctamente.' }
 	  		else
 	  			format.html { render :edit }
@@ -37,10 +36,29 @@ class ChofersController < ApplicationController
 	  	end
 	 end
 	
+	
 	def destroy
-		chofer = Chofer.find(params[:id])
-		chofer.destroy
-		redirect_to chofers_path
-	end
+	    
+	    chofer = Chofer.find(params[:id])
+	    if Trip.tiene_chofer(params[:id]).exists?
+	    	respond_to do |format|
+	    		format.html {redirect_to chofers_path, alert: 'El chofer no pudo eliminarse ya que tiene un viaje asociado activo.' }
+	    	end
+	    else
+	    	chofer.destroy
+	      	redirect_to chofers_path, notice: 'El chofer se eliminó correctamente.' 
+	    	
+		end
+  	end
+
+
+	private
+		def set_chofer
+	    	@chofer = Chofer.find(params[:id])
+	    end
+	    def chofer_params
+			params.require(:chofer).permit(:nombre, :apellido, :dni, :inicio_actividad)	
+	    end	
+
 
   	end
